converting to typescript
