--- src/functions/index.js
@@ -1,3 +1,0 @@
-export * from './object';
-export * from './schema';
-export * from './DBML';
