--- src/functions/object.js
@@ -1,19 +1,0 @@
-export const checkSame = (a1, a2) => {
-  if (typeof a1 !== typeof a2) return false;
-  if (typeof a1 === 'object') return checkSameObject(a1, a2);
-  return a1 === a2;
-};
-
-export const checkSameObject = (obj1, obj2) => {
-  if (typeof obj1 !== typeof obj2) return false;
-  let key1, key2;
-  try {
-    key1 = Object.keys(obj1);
-    key2 = Object.keys(obj2);
-  } catch (e) {
-    return obj1 === obj2;
-  }
-  if (key1.length !== key2.length) return false;
-  if (!key1.every((t) => key2.includes(t))) return false;
-  return key1.every((t) => checkSame(obj1[t], obj2[t]));
-};
