--- src/functions/schema/index.js
@@ -1,2 +1,0 @@
-export * from './sync';
-export * from './schema';
