+++ src/functions/schema/index.ts
@@ -1,0 +1,2 @@
+export * from './sync';
+export * from './schema';
