--- src/index.js
@@ -1,6 +1,0 @@
-export * from './constants';
-export * from './functions';
-
-import {ParseField, Indexes, ParseClassSchema} from '@Types/fields';
-
-export {ParseClassSchema, ParseField, Indexes};
